refactor(phone-bridge): centralize pip bridge policy
