+++ src/agent/internal/agent/tools_phone_bridge.go
@@ -245,46 +245,3 @@ func (s *ToolSet) RegisterPhoneBridge(bridge *PhoneBridge) {
 	s.tools["contacts"] = NewContactsTool(bridge, s.phoneBridgeRestorer)
 	s.tools["notification"] = NewNotificationTool(bridge, s.phoneBridgeRestorer)
 }
-
-func isPhoneBridgeToolName(name string) bool {
-	switch name {
-	case "open_app", "clipboard", "calendar", "contacts", "notification":
-		return true
-	default:
-		return false
-	}
-}
-
-func phoneBridgeToolAvailable(status PhoneBridgeStatus, name string) bool {
-	switch name {
-	case "open_app":
-		return !phoneBridgePiPBackgroundEnabled(status)
-	case "clipboard", "calendar", "contacts", "notification":
-		return status.Connected || phoneBridgeCanUsePiPBackground(status, phoneBridgeBackgroundCommandTypeForTool(name))
-	default:
-		return true
-	}
-}
-
-func phoneBridgePiPBackgroundEnabled(status PhoneBridgeStatus) bool {
-	if status.PipBridgeEnabled == nil || !*status.PipBridgeEnabled {
-		return false
-	}
-	state := strings.ToLower(strings.TrimSpace(status.AppState))
-	return state == "" || state == "background" || state == "inactive"
-}
-
-func phoneBridgeBackgroundCommandTypeForTool(name string) string {
-	switch name {
-	case "clipboard":
-		return "clipboard_read"
-	case "calendar":
-		return "calendar_query"
-	case "contacts":
-		return "contacts_query"
-	case "notification":
-		return "notification_send"
-	default:
-		return ""
-	}
-}
